requirements.txt
langchain
langchain_openai
langchain_core
langchain_community
cassio
datasets
tiktoken
PyPDF2
python_dotenv
streamlit